+++ core/src/main/java/org/hisp/dhis/android/sdk/synchronization/data/event/EventRemoteDataSource.java
@@ -48,7 +48,6 @@ public class EventRemoteDataSource {
     }
 
     private ImportSummary putEvent(Event event, DhisApi dhisApi) throws APIException {
-
         Response response = dhisApi.putEvent(event.getEvent(), event);
 
         return getImportSummary(response);
