+++ core/src/main/java/org/hisp/dhis/android/sdk/synchronization/domain/event/EventSynchronizer.java
@@ -8,16 +8,17 @@ import org.hisp.dhis.android.sdk.network.APIException;
 import org.hisp.dhis.android.sdk.persistence.models.Event;
 import org.hisp.dhis.android.sdk.persistence.models.FailedItem;
 import org.hisp.dhis.android.sdk.persistence.models.ImportSummary;
+import org.hisp.dhis.android.sdk.synchronization.domain.common.Synchronizer;
 import org.hisp.dhis.android.sdk.synchronization.domain.faileditem.IFailedItemRepository;
 
-public class EventSynchronizer {
-    //coordinate one type of item
-
+public class EventSynchronizer extends Synchronizer{
     IEventRepository mEventRepository;
     IFailedItemRepository mFailedItemRepository;
 
     public EventSynchronizer(IEventRepository eventRepository,
             IFailedItemRepository failedItemRepository) {
+        super(failedItemRepository);
+
         mEventRepository = eventRepository;
         mFailedItemRepository = failedItemRepository;
     }
@@ -31,13 +32,13 @@ public class EventSynchronizer {
 
                 event.setFromServer(true);
                 mEventRepository.save(event);
-                mFailedItemRepository.clearFailedItem(EVENT, event.getLocalId());
+                super.clearFailedItem(EVENT, event.getLocalId());
 
             } else if (ImportSummary.ERROR.equals(importSummary.getStatus())) {
-                mFailedItemRepository.handleImportSummaryError(importSummary, EVENT, 200, id);
+                super.handleImportSummaryError(importSummary, EVENT, 200, id);
             }
         } catch (APIException api) {
-            mFailedItemRepository.handleSerializableItemException(api, FailedItem.EVENT,
+            super.handleSerializableItemException(api, FailedItem.EVENT,
                     event.getLocalId());
         }
     }
